+++ src/style/mod.rs
@@ -32,9 +32,30 @@ pub enum AutoStyleMode {
     None,
     /// Include auto-styles as part of element `<style>` attributes.
     Inline,
+    /// As per `inline`, but avoid 'CDATA' to generate valid HTML
+    InlineHtml,
     /// Generate CSS auto-style rules in a separate `<style>` element.
     #[default]
     Css,
+    /// As per `css`, but avoid 'CDATA' to generate valid HTML
+    CssHtml,
+}
+
+impl AutoStyleMode {
+    /// Returns `true` if this mode should use CDATA sections for inline styles.
+    pub fn use_cdata(&self) -> bool {
+        !matches!(self, AutoStyleMode::InlineHtml | AutoStyleMode::CssHtml)
+    }
+
+    pub fn variants() -> Vec<String> {
+        vec![
+            AutoStyleMode::None.to_string(),
+            AutoStyleMode::Inline.to_string(),
+            AutoStyleMode::InlineHtml.to_string(),
+            AutoStyleMode::Css.to_string(),
+            AutoStyleMode::CssHtml.to_string(),
+        ]
+    }
 }
 
 impl std::str::FromStr for AutoStyleMode {
@@ -44,7 +65,9 @@ impl std::str::FromStr for AutoStyleMode {
         match s {
             s if s == AutoStyleMode::None.to_string() => Ok(AutoStyleMode::None),
             s if s == AutoStyleMode::Inline.to_string() => Ok(AutoStyleMode::Inline),
+            s if s == AutoStyleMode::InlineHtml.to_string() => Ok(AutoStyleMode::InlineHtml),
             s if s == AutoStyleMode::Css.to_string() => Ok(AutoStyleMode::Css),
+            s if s == AutoStyleMode::CssHtml.to_string() => Ok(AutoStyleMode::CssHtml),
             _ => Err(Error::InvalidValue("auto-style-mode".into(), s.into())),
         }
     }
@@ -55,7 +78,9 @@ impl std::fmt::Display for AutoStyleMode {
         let s = match self {
             AutoStyleMode::None => "none",
             AutoStyleMode::Inline => "inline",
+            AutoStyleMode::InlineHtml => "inline-html",
             AutoStyleMode::Css => "css",
+            AutoStyleMode::CssHtml => "css-html",
         };
         f.write_str(s)
     }